Made breaking changes making implementation explicit
